requirements.txt
# For REBEL/ReFinEd:
boto3
kglab
openai
pandas
pypdf
transformers
torch
https://github.com/amazon-science/ReFinED/archive/refs/tags/V1.zip

# For Claude:
anthropic
python-dotenv
pathlib
httpx